Move Server to a separate mod
